Extract a common TestingSystemAccessControlContext
